backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.200
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3